+++ tools/pal2rgb.c
@@ -1,4 +1,4 @@
-/* $Id: pal2rgb.c,v 1.7 2004-09-02 14:46:27 dron Exp $ */
+/* $Id: pal2rgb.c,v 1.8 2004-09-21 12:20:04 dron Exp $ */
 
 /*
  * Copyright (c) 1988-1997 Sam Leffler
@@ -231,7 +231,7 @@ processCompressOptions(char* opt)
 		compression = COMPRESSION_PACKBITS;
 	else if (strneq(opt, "jpeg", 4)) {
 		char* cp = strchr(opt, ':');
-		if (cp && isdigit(cp[1]))
+		if (cp && isdigit((int)cp[1]))
 			quality = atoi(cp+1);
 		if (cp && strchr(cp, 'r'))
 			jpegcolormode = JPEGCOLORMODE_RAW;
